# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_web_controllers_http.SwaggerCreateWorkOrderRequest:
    properties:
      dueDateUtc:
        type: string
      notes:
        type: string
      responsibleUserId:
        type: string
      saleId:
        type: string
      status:
        type: string
    required:
    - saleId
    type: object
  internal_web_controllers_http.SwaggerCreateWorkerRequest:
    properties:
      email:
        type: string
      fixedSalary:
        type: number
      fullName:
        type: string
      notes:
        type: string
      phone:
        type: string
      salaryPeriod:
        type: string
    required:
    - fullName
    type: object
...
  internal_web_controllers_http.SwaggerErrorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          type: string
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  internal_web_controllers_http.SwaggerMeta:
    properties:
      total:
        type: integer
    type: object
  internal_web_controllers_http.SwaggerOKFlag:
    properties:
      ok:
        type: boolean
    type: object
  internal_web_controllers_http.SwaggerOKFlagResponse:
    properties:
      data:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlag'
    type: object
...
  internal_web_controllers_http.SwaggerPayCommissionRequest:
    properties:
      amount:
        type: number
      method:
        type: string
      notes:
        type: string
      workerId:
        type: string
    required:
    - amount
    - method
    - workerId
    type: object
  internal_web_controllers_http.SwaggerPaySalaryRequest:
    properties:
      amount:
        type: number
      method:
        type: string
      notes:
        type: string
      workerId:
        type: string
    required:
    - amount
    - method
    - workerId
    type: object
...
  internal_web_controllers_http.SwaggerSetSettingRequest:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  internal_web_controllers_http.SwaggerSettingResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.AppSetting'
    type: object
...
  internal_web_controllers_http.SwaggerUpdateWorkOrderRequest:
    properties:
      dueDateUtc:
        type: string
      notes:
        type: string
      responsibleUserId:
        type: string
      saleId:
        type: string
      status:
        type: string
    required:
    - saleId
    type: object
  internal_web_controllers_http.SwaggerUpdateWorkerRequest:
    properties:
      email:
        type: string
      fixedSalary:
        type: number
      fullName:
        type: string
      notes:
        type: string
      phone:
        type: string
      salaryPeriod:
        type: string
    required:
    - fullName
    type: object
...
  internal_web_controllers_http.SwaggerWorkOrderListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.WorkOrder'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerWorkOrderResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.WorkOrder'
    type: object
  internal_web_controllers_http.SwaggerWorkerListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Worker'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerWorkerResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Worker'
    type: object
...
  photogallery_api_go_internal_domain_entities.AppSetting:
    properties:
      key:
        type: string
      updatedAtUtc:
        type: string
      updatedByUserID:
        type: string
      value:
        type: string
    type: object
...
  photogallery_api_go_internal_domain_entities.WorkOrder:
    properties:
      createdAtUtc:
        type: string
      deliveredAtUtc:
        type: string
      dueDateUtc:
        type: string
      id:
        type: string
      notes:
        type: string
      responsibleUserID:
        type: string
      saleID:
        type: string
      status:
        type: string
      updatedAtUtc:
        type: string
    type: object
  photogallery_api_go_internal_domain_entities.Worker:
    properties:
      createdAtUtc:
        type: string
      email:
        type: string
      fixedSalary:
        format: float64
        type: number
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      notes:
        type: string
      phone:
        type: string
      salaryPeriod:
        type: string
      updatedAtUtc:
        type: string
    type: object
info:
  contact: {}
  description: API REST v1 Foto-tienda
  title: Photo Gallery API
  version: "1.0"
paths:
...
  /api/v1/settings/:key:
    get:
      description: Obtiene una configuración del sistema por clave.
      parameters:
      - description: Setting key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerSettingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetSystemSetting
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Actualiza el valor de una configuración del sistema.
      parameters:
      - description: Set setting payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerSetSettingRequest'
      - description: Setting key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: SetSystemSetting
      tags:
      - settings
...
  /api/v1/work-orders:
    get:
      description: Lista paginada de órdenes de trabajo con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkOrderListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListWorkOrders
      tags:
      - work-orders
    post:
      consumes:
      - application/json
      description: Crea una nueva orden de trabajo.
      parameters:
      - description: Create work order payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateWorkOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateWorkOrder
      tags:
      - work-orders
  /api/v1/work-orders/:id:
    get:
      description: Obtiene una orden de trabajo por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetWorkOrder
      tags:
      - work-orders
    put:
      consumes:
      - application/json
      description: Actualiza una orden de trabajo existente.
      parameters:
      - description: Update work order payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateWorkOrderRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateWorkOrder
      tags:
      - work-orders
  /api/v1/workers:
    get:
      description: Lista paginada de trabajadores con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkerListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListWorkers
      tags:
      - workers
    post:
      consumes:
      - application/json
      description: Crea un nuevo trabajador.
      parameters:
      - description: Create worker payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateWorkerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateWorker
      tags:
      - workers
  /api/v1/workers/:id:
    get:
      description: Obtiene un trabajador por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetWorker
      tags:
      - workers
    put:
      consumes:
      - application/json
      description: Actualiza un trabajador existente.
      parameters:
      - description: Update worker payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateWorkerRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerWorkerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateWorker
      tags:
      - workers
  /api/v1/workers/:id/activate:
    patch:
      description: Activa un trabajador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ActivateWorker
      tags:
      - workers
  /api/v1/workers/:id/deactivate:
    patch:
      description: Desactiva un trabajador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeactivateWorker
      tags:
      - workers
  /api/v1/workers/pay-commissions:
    post:
      consumes:
      - application/json
      description: Ejecuta el pago de comisiones pendientes.
      parameters:
      - description: Pay commissions payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerPayCommissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: PayCommissions
      tags:
      - workers
  /api/v1/workers/pay-salary:
    post:
      consumes:
      - application/json
      description: Ejecuta el pago de salario.
      parameters:
      - description: Pay salary payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerPaySalaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: PaySalary
      tags:
      - workers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"